requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-language==2.0.0
google-cloud-speech==2.0.1
google-cloud-texttospeech==2.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
libcst==0.3.16
mypy-extensions==0.4.3
oauth2client==4.1.3
pip-autoremove==0.9.1
playsound==1.2.2
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pygame==2.0.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
SpeechRecognition==3.8.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
vext==0.7.4
vext.gi==0.7.4
